+++ src/net-protocol/export/ProtocolWrangler.js
@@ -1,11 +1,13 @@
 // Copyright 2022-2024 the Lactoserv Authors (Dan Bornstein et alia).
 // SPDX-License-Identifier: Apache-2.0
 
+import { inspect } from 'node:util';
+
 import { Threadlet } from '@this/async';
 import { ProductInfo } from '@this/host';
 import { IntfLogger } from '@this/loggy-intf';
 import { IncomingRequest, IntfRequestHandler, OutgoingResponse, RequestContext,
-  TypeNodeRequest, TypeNodeResponse }
+  StatusResponse, TypeNodeRequest, TypeNodeResponse }
   from '@this/net-util';
 import { Methods, MustBe } from '@this/typey';
 
@@ -400,25 +402,24 @@ export class ProtocolWrangler {
       return OutgoingResponse.makeMetaResponse(400); // "Bad Request."
     }
 
-    try {
-      const result = await this.#requestHandler.handleRequest(request, null);
+    const result = await this.#requestHandler.handleRequest(request, null);
 
-      if (result instanceof OutgoingResponse) {
-        return result;
-      } else if (result === null) {
-        // The configured `requestHandler` didn't actually handle the request.
-        // Respond with a vanilla `404` error. (If the client wants something
-        // fancier, they can do it themselves.)
-        const bodyExtra = request.urlForLog;
-        return OutgoingResponse.makeNotFound({ bodyExtra });
-      } else {
-        // Caught by our direct caller, `#respondToRequest()`.
-        throw new Error(`Strange result from \`handleRequest\`: ${result}`);
-      }
-    } catch (e) {
-      // `500` == "Internal Server Error."
-      const bodyExtra = e.stack ?? e.message ?? '<unknown>';
-      return OutgoingResponse.makeMetaResponse(500, { bodyExtra });
+    if (result instanceof OutgoingResponse) {
+      return result;
+    } else if (result instanceof StatusResponse) {
+      return result.responseFor(request);
+    } else if (result === null) {
+      // The configured `requestHandler` didn't actually handle the request.
+      // Respond with a simple "not found". (If the client wants something
+      // fancier, they can do it themselves.)
+      return StatusResponse.NOT_FOUND.responseFor(request);
+    } else {
+      // The error thrown here is caught by our direct caller,
+      // `#respondToRequest()`.
+      const type      = typeof result;
+      const inspected = inspect(result);
+      throw new Error(
+        `Strange result (type \`${type}\` from \`handleRequest\`:\n${inspected}\n`);
     }
   }
 
@@ -458,7 +459,7 @@ export class ProtocolWrangler {
       result ??= await this.#handleRequest(request, outerContext);
     } catch (e) {
       // `500` == "Internal Server Error."
-      const bodyExtra = e.stack ?? e.message ?? '<unknown error>';
+      const bodyExtra = e.stack ?? e.message ?? '<unknown>';
       result = OutgoingResponse.makeMetaResponse(500, { bodyExtra });
     }