+++ src/net-util/export/TypeResponse.js
@@ -2,6 +2,7 @@
 // SPDX-License-Identifier: Apache-2.0
 
 import { OutgoingResponse } from '#x/OutgoingResponse';
+import { StatusResponse } from '#x/StatusResponse';
 
 
 /**
@@ -9,9 +10,12 @@ import { OutgoingResponse } from '#x/OutgoingResponse';
  * network requests, when using the classes in this module.
  *
  * * {@link OutgoingResponse} -- A complete response.
+ * * {@link StatusResponse} -- A response which _just_ notes the status code.
+ *   Code "near" the protocol handler is expected to expand these into complete
+ *   responses when appropriate.
  * * `null` -- Indicator that the request was not handled (which is different
  *   than being handled with a result of "not found").
  *
- * @typedef {OutgoingResponse|null} TypeResponse
+ * @typedef {OutgoingResponse|StatusResponse|null} TypeResponse
  */
 export const TypeResponse = Symbol('TypeResponse');